requirements.txt
aiosqlite==0.22.1
bcrypt
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pydantic-settings==2.7.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
asyncpg==0.31.0
certifi==2026.1.4
cffi==2.0.0
circuitbreaker==2.1.3
click==8.0.4
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
functions==0.7.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==0.10.0
numpy==2.4.0
oci==2.164.2
oci-cli==3.71.4
packaging==25.0
pluggy==1.6.0
prompt-toolkit==3.0.43
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyOpenSSL==25.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
schemas==0.7.1
setuptools==80.9.0
six==1.17.0
spglib==2.7.0
SQLAlchemy==2.0.45
starlette==0.50.0
terminaltables==3.1.10
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
wcwidth==0.2.14
